server/requirements.txt
curl-cffi==0.6.2
fake-useragent==1.5.1
importlib-metadata==7.1.0
ipykernel==6.29.3
openai==1.14.2
pickleshare==0.7.5
pip-chill==1.0.3
pysocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
rich==13.7.1
trio-websocket==0.11.1
